+++ RELEASES.md
@@ -10,3 +10,13 @@ Based on:
 - [typescript v0.0.1-alpha.1] .
 ### Releases
 - [NPM v0.0.1-alpha.1] https://www.npmjs.com/package/@avalabs/avacloud-sdk/v/0.0.1-alpha.1 - .
+
+## 2024-08-28 19:36:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.382.0 (2.404.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.0.1-alpha.2] .
+### Releases
+- [NPM v0.0.1-alpha.2] https://www.npmjs.com/package/@avalabs/avacloud-sdk/v/0.0.1-alpha.2 - .